chore: Renaming source code dir in cli & prerender-data-provider
