+++ packages/cli/package.json
@@ -18,14 +18,14 @@
   },
   "license": "MIT",
   "author": "The Preact CLI Authors (https://github.com/preactjs/preact-cli/contributors)",
-  "main": "lib/index.js",
+  "main": "src/index.js",
   "types": "types.d.ts",
   "bin": {
-    "preact": "lib/index.js"
+    "preact": "src/index.js"
   },
   "files": [
     "sw",
-    "lib",
+    "src",
     "babel",
     "types.d.ts"
   ],
