+++ packages/cli/src/commands/build.js
