+++ packages/cli/src/commands/index.js
