+++ packages/cli/src/commands/list.js
