+++ packages/cli/src/commands/validate-args.js
