+++ packages/cli/src/commands/watch.js
