+++ packages/cli/src/index.js
