+++ packages/cli/src/lib/babel-config.js
