+++ packages/cli/src/lib/entry.js
