+++ packages/cli/src/lib/logo.js
