+++ packages/cli/src/lib/setup.js
