+++ packages/cli/src/resources/body-end.ejs
