+++ packages/cli/src/resources/favicon.ico
