+++ packages/cli/src/resources/head-end.ejs
