+++ packages/cli/src/resources/icon.png
