+++ packages/cli/src/resources/manifest.json
