+++ packages/cli/src/resources/static-app.json
