+++ packages/cli/src/resources/sw-debug.js
