+++ packages/cli/src/resources/template.html
