+++ packages/cli/src/symbols.js
