+++ packages/cli/tests/images/build.js
@@ -1,6 +1,6 @@
 const common = {
-	'polyfills.70f66.js': 6426,
-	'polyfills.70f66.js.map': 21668,
+	'polyfills.1ea90.js': 6426,
+	'polyfills.1ea90.js.map': 21668,
 };
 
 exports.default = Object.assign({}, common, {
@@ -17,8 +17,8 @@ exports.default = Object.assign({}, common, {
 	'ssr-build/ssr-bundle.js.map': 32557,
 	'ssr-build/asset-manifest.json': 178,
 	'bundle.259c5.css': 901,
-	'bundle.fbf55.js': 21429,
-	'bundle.fbf55.js.map': 111801,
+	'bundle.fcca3.js': 21429,
+	'bundle.fcca3.js.map': 111801,
 	'favicon.ico': 15086,
 	'index.html': 2034,
 	'manifest.json': 455,
