+++ packages/prerender-data-provider/package.json
@@ -15,9 +15,9 @@
   },
   "license": "MIT",
   "author": "The Preact CLI Authors (https://github.com/preactjs/preact-cli/contributors)",
-  "main": "lib/index.js",
+  "main": "src/index.js",
   "files": [
-    "lib"
+    "src"
   ],
   "peerDependencies": {
     "preact": "^10.0.0"
